Solve the problem
Problem is I simply return the result from a recursive function. An d
this will always lead to return nothing. It's hard to find. I use a
globel value to store the search result. I know this is not the best
solution. But I don't want return the result all the way back to the
first call function.